% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlockCov.R
\docType{package}
\name{BlockCov}
\alias{BlockCov}
\alias{BlockCov-package}
\title{\code{BlockCov} package}
\description{
Estimation of Large Block Covariance Matrices
}
\details{
See the README on
\href{https://cran.r-project.org/package=BlockCov/vignettes/Vignettes.pdf}{CRAN}
\href{https://github.com/Marie-PerrotDockes/BlockCov#readme}{GitHub}
}
