% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_pa.R
\name{PA}
\alias{PA}
\title{Title}
\usage{
PA(E, times = 10)
}
\arguments{
\item{E}{the observation matrix such that each of its row has a block structure correlation matrix Sigma wich has a low rank once its diagonal is removed.}

\item{times}{number of random sampling}
}
\value{
the mean of the eigen values of the \code{times} sampled matrix
}
\description{
Title
}
\examples{
n <- 30
q <- 100
Sigma <- Simu_Sigma(q = q, diag = FALSE, equal = TRUE)
Matrix::image(Sigma)
E <- matrix(rnorm(n * q), ncol = q) \%*\% chol(as.matrix(Sigma))
random_eigen <- PA(E, times = 10)
}
