% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simu_Sigma.R
\name{Simu_Sigma}
\alias{Simu_Sigma}
\title{This function generates a block structured symmetric positive definite matrix to test the BlockCov methodology.}
\usage{
Simu_Sigma(q, diag = TRUE, equal = TRUE)
}
\arguments{
\item{q}{integer corresponding to the size of the covariance matrix.}

\item{diag}{logical, whether or not the covariance matrix is block-diagonal.}

\item{equal}{logical, whether or not the values in the blocks are equal.}
}
\value{
Sigma a correlation matrix to test the BlockCov methodology.
}
\description{
This function generates a block structured symmetric positive definite matrix to test the BlockCov methodology.
}
\examples{
Sigma <- Simu_Sigma(q = 100, diag = FALSE, equal = TRUE)
Matrix::image(Sigma)
}
\references{
M. Perrot-Dockès and C. Lévy-Leduc. "Estimation of large block covariance matrices: Application to the analysis of gene expression data", arXiv:1806.10093.
}
